persistence/requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.28.0
